+++ modules/api.js
@@ -1,6 +1,6 @@
 import { currentDate } from "./utils.js";
 
-export let token = null;
+export let token = 'Bearer asb4c4boc86gasb4c4boc86g37w3cc3bo3b83k4g37k3bk3cg3c03ck4k';
 export const getToken = () => {
   return `Bearer ${token}`;
 };
