+++ modules/validation.js
@@ -1,3 +1,4 @@
+import { buttonElement } from "./renderOptional.js";
 export const trimValue = (element) => element.value.trim();
 
 export const setError = (element, message) => {};
